Move `samsungtvws` to Root Directory
