--- custom_components/samsungtv_custom/samsungtvws/__init__.py
@@ -1,23 +1,0 @@
-"""
-SamsungTVWS - Samsung Smart TV WS API wrapper
-
-Copyright (C) 2019 Xchwarze
-
-    This library is free software; you can redistribute it and/or
-    modify it under the terms of the GNU Lesser General Public
-    License as published by the Free Software Foundation; either
-    version 2.1 of the License, or (at your option) any later version.
-
-    This library is distributed in the hope that it will be useful,
-    but WITHOUT ANY WARRANTY; without even the implied warranty of
-    MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
-    Lesser General Public License for more details.
-
-    You should have received a copy of the GNU Lesser General Public
-    License along with this library; if not, write to the Free Software
-    Foundation, Inc., 51 Franklin Street, Fifth Floor,
-    Boston, MA  02110-1335  USA
-
-"""
-from .remote import SamsungTVWS
-from .shortcuts import SamsungTVShortcuts
